# crates/paladin-web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Paladin Agent API",
    "description": "HTTP API for executing and managing resident Paladin agents. Agent routes are served under `/v1`; `/health`, `/ready`, and the docs are unversioned.",
    "contact": {
      "name": "Juha Kukkonen",
      "email": "juha7kukkonen@gmail.com"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.7.0"
  },
  "paths": {
    "/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "`GET /agents` — list every registered agent as a safe [`AgentSummary`].",
        "description": "Always returns `200 OK` with a JSON array (empty when no agents are registered).\nOrder is unspecified.",
        "operationId": "list_agents",
        "responses": {
          "200": {
            "description": "Registered agents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AgentSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "bearer_token": []
          }
        ]
      },
      "post": {
        "tags": [
          "agents"
        ],
        "summary": "`POST /agents` — register a new agent at runtime from an [`AgentSpec`].",
        "description": "Because `paladin-web` cannot build a [`Paladin`] itself, this delegates to the\ninjected [`AgentProvisioner`]. Returns:\n- `201 Created` with the new agent's [`AgentSummary`] on success;\n- `409 Conflict` if an agent is already registered under the spec's id;\n- `422 Unprocessable Entity` if provisioning fails;\n- `400 Bad Request` (via the `Json` extractor) if the body is missing/invalid;\n- `501 Not Implemented` if no provisioner is wired (registration disabled).",
        "operationId": "register_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSpec"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Agent registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSummary"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "403": {
            "description": "Admin role required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Agent id already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Provisioning failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "501": {
            "description": "Runtime registration not enabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "api_key": []
          },
          {
            "bearer_token": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentSpec": {
        "type": "object",
        "description": "Declarative description of an agent to provision at runtime.\n\nSent in the body of `POST /agents`. Because `paladin-web` cannot itself build a\n[`Paladin`] (that needs an `LlmPort` and the builder, which live behind the\nfacade), an [`AgentProvisioner`] turns this spec into a concrete\n`(Paladin, executor)` pair.\n\nOnly the fields needed to identify and shape an agent are included here; the\nprovisioner is free to apply defaults for everything else. More fields can be\nadded without breaking existing callers (all optional fields use `#[serde(default)]`).",
        "required": [
          "id",
          "name",
          "model",
          "system_prompt"
        ],
        "properties": {
          "allowed_roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Roles permitted to invoke this agent (`\"admin\"`/`\"user\"`); empty ⇒ any caller."
          },
          "id": {
            "type": "string",
            "description": "Registry id (the path segment in `/agents/{id}/…`). Client-supplied and unique."
          },
          "model": {
            "type": "string",
            "description": "LLM model identifier (e.g. `\"gpt-4\"`)."
          },
          "name": {
            "type": "string",
            "description": "Human-friendly display name."
          },
          "stop_words": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional stop words that terminate execution."
          },
          "system_prompt": {
            "type": "string",
            "description": "System prompt defining the agent's behavior."
          },
          "temperature": {
            "type": [
              "number",
              "null"
            ],
            "format": "float",
            "description": "Optional response randomness (`0.0`–`1.0`); the provisioner applies a default when absent."
          },
          "timeout_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Optional per-agent execution timeout (seconds); `None` uses the server default.",
            "minimum": 0
          }
        }
      },
      "AgentSummary": {
        "type": "object",
        "description": "Safe, public-facing summary of an agent for the discovery endpoints.\n\nDeliberately omits anything sensitive. Note that secrets (API keys, provider\ncredentials) never live on the [`Paladin`] entity in the first place — they are\nsupplied to executors at composition time — so none can leak here. The raw system\nprompt is reduced to a short `description` preview rather than returned verbatim\n(see PRD Open Question 1 on whether to omit it entirely).",
        "required": [
          "id",
          "name",
          "model",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "Short, single-line preview derived from the system prompt."
          },
          "id": {
            "type": "string",
            "description": "Registry id (the `{id}` path segment)."
          },
          "model": {
            "type": "string",
            "description": "LLM model identifier."
          },
          "name": {
            "type": "string",
            "description": "Human-friendly display name."
          }
        }
      },
      "ApiErrorBody": {
        "type": "object",
        "description": "OpenAPI schema for the error envelope returned by every failing handler.\n\nMirrors [`ApiError::to_body`] (`{ \"error\": { \"code\", \"message\", \"details\" } }`); it exists\npurely so the generated spec can describe error responses. The runtime body is built by\n`ApiError`, not this type.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail",
            "description": "The error detail object."
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "description": "The `error` object inside [`ApiErrorBody`].",
        "required": [
          "code",
          "message",
          "details"
        ],
        "properties": {
          "code": {
            "type": "string",
            "description": "Stable, machine-readable error code (e.g. `\"not_found\"`, `\"unauthorized\"`)."
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "description": "Optional structured context; `null` when absent."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message."
          }
        }
      },
...
    },
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      },
      "bearer_token": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}